Input Viewer 구현
