+++ .gitignore
@@ -19,3 +19,4 @@ bin/
 *.iws
 *.iml
 *.ipr
+/lib/
