+++ production/step3/main/java/racingCar/README.md
@@ -1,0 +1,18 @@
+#java-racingcar
+##3단계 - 자동차 경주
+>기능 요구사항
+>>초간단 자동차 경주 게임을 구현한다.
++주어진 횟수 동안 n대의 자동차는 전진 또는 멈출 수 있다.
++사용자는 몇 대의 자동차로 몇 번의 이동을 할 것인지를 입력할 수 있어야 한다.
++전진하는 조건은 0에서 9 사이에서 random 값을 구한 후 random 값이 4이상일 경우이다.
++자동차의 상태를 화면에 출력한다. 어느 시점에 출력할 것인지에 대한 제약은 없다.
+- - - 
+>기능 목록
+>>-view
++   자동차 대수 입력창 구현
++   시도할 회수 입력창 구현
++   실행결과 결과창 구현
+>>-model
++ 자동차클래스 구현(위치,상태,각각 getter와setter,초기화)
+>>-controller
++ 1개 라운드 게임진행 기능 구현
