+++ step3/src/main/java/racingCar/controller/Game.java
@@ -1,13 +1,4 @@
-package src.main.java.racingCar.controller;
-
-import src.main.java.racingCar.view.InputViewer;
+package main.java.racingCar.controller;
 
 public class Game {
-    public static void main(){
-        System.out.println("--");
-        InputViewer inputViewer=new InputViewer();
-        int carNumber = inputViewer.carInputViewer();
-        int gameNumber = inputViewer.tryInputViewer();
-
-    }
 }
