+++ step3/src/main/java/racingCar/controller/RacingGame.java
@@ -1,0 +1,15 @@
+package main.java.racingCar.controller;
+
+import main.java.racingCar.view.InputViewer;
+
+public class Game {
+    Game(){}
+    public static void main(){
+        System.out.println("--");
+        InputViewer inputViewer=new InputViewer();
+        int carNumber = inputViewer.carInputViewer();
+        int gameNumber = inputViewer.tryInputViewer();
+        Game game = new Game();
+        
+    }
+}
