+++ step3/src/main/java/racingCar/view/OutputViewer.java
@@ -1,4 +1,4 @@
-package src.main.java.racingCar.view;
+package main.java.racingCar.view;
 
 public class OutputViewer {
     void resultOutputViewer(){}
