+++ step3/test/main/java/racingCar/controller/Main.java
@@ -1,3 +1,5 @@
+package main.java.racingCar.controller;
+
 public class Main {
     public static void main(String[] args) {
         System.out.println("---");
